#155 Delete Bookmark ViewModel
