+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/di/appModules.kt
@@ -7,7 +7,6 @@ val appModules = arrayListOf(
     , detailModule
     , apiModule
     , mapModule
-    , bookmarkModule
     , searchModule
     , reverseGeoApiModule
     , homeModule
