+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/di/homeModule.kt
@@ -1,10 +1,12 @@
 package kr.co.connect.boostcamp.livewhere.di
 
+import kr.co.connect.boostcamp.livewhere.repository.BookmarkRepositoryImpl
 import kr.co.connect.boostcamp.livewhere.ui.main.HomeViewModel
 import org.koin.androidx.viewmodel.ext.koin.viewModel
 import org.koin.dsl.module.module
 
 val homeModule = module {
+    factory("bookmarkRepository") { BookmarkRepositoryImpl(get("bookmarkDAO")) }
 
-    viewModel { HomeViewModel() }
+    viewModel { HomeViewModel(get("bookmarkRepository")) }
 }
