--- app/src/main/java/kr/co/connect/boostcamp/livewhere/ui/main/BookmarkViewModel.kt
@@ -1,50 +1,0 @@
-package kr.co.connect.boostcamp.livewhere.ui.main
-
-import android.annotation.SuppressLint
-import android.view.View
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import io.reactivex.android.schedulers.AndroidSchedulers
-import io.reactivex.schedulers.Schedulers
-import kr.co.connect.boostcamp.livewhere.data.entity.BookmarkEntity
-import kr.co.connect.boostcamp.livewhere.repository.BookmarkRepositoryImpl
-import kr.co.connect.boostcamp.livewhere.ui.BaseViewModel
-
-class BookmarkViewModel(private val bookmarkRepositoryImpl: BookmarkRepositoryImpl) : BaseViewModel() {
-    var isEmptyBookmark: Int = View.VISIBLE
-
-    private val _bookmarkEntity = MutableLiveData<List<BookmarkEntity>>()
-    val bookmarkEntity: LiveData<List<BookmarkEntity>>
-        get() = _bookmarkEntity
-
-    private val _sendAddress = MutableLiveData<String>()
-    val sendAddress: LiveData<String>
-        get() = _sendAddress
-
-    init {
-        getBookmark()
-    }
-
-    @SuppressLint("CheckResult")
-    fun getBookmark() {
-        getCompositeDisposable().add(
-            bookmarkRepositoryImpl.getBookmark()
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .subscribe({
-                    _bookmarkEntity.postValue(it)
-                }, {
-
-                })
-        )
-        if(bookmarkEntity.value.isNullOrEmpty()) {
-            isEmptyBookmark = View.VISIBLE
-        } else {
-            isEmptyBookmark = View.GONE
-        }
-    }
-
-    fun setSendText(text: String) {
-        _sendAddress.postValue(text)
-    }
-}
