+++ app/src/main/java/kr/co/connect/boostcamp/livewhere/ui/main/adapter/BookmarkRecyclerViewAdapter.kt
@@ -8,11 +8,11 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import kr.co.connect.boostcamp.livewhere.data.entity.BookmarkEntity
 import kr.co.connect.boostcamp.livewhere.databinding.ItemBookmarkRecyclerViewBinding
-import kr.co.connect.boostcamp.livewhere.ui.main.BookmarkViewModel
+import kr.co.connect.boostcamp.livewhere.ui.main.HomeViewModel
 
 class BookmarkRecyclerViewAdapter(
     private val lifecycleOwner: LifecycleOwner,
-    private val bookmarkViewModel: BookmarkViewModel
+    private val homeViewModel: HomeViewModel
 ) : RecyclerView.Adapter<BookmarkRecyclerViewAdapter.BookmarkViewHolder>() {
 
     private var list = listOf<BookmarkEntity>()
@@ -31,19 +31,18 @@ class BookmarkRecyclerViewAdapter(
     override fun getItemCount() = list.size
 
     override fun onBindViewHolder(holder: BookmarkViewHolder, position: Int) {
-        holder.bind(lifecycleOwner, list[position], bookmarkViewModel)
+        holder.bind(lifecycleOwner, list[position], homeViewModel)
     }
 
     inner class BookmarkViewHolder(
         private val itemBinding: ItemBookmarkRecyclerViewBinding
     ) : RecyclerView.ViewHolder(itemBinding.root) {
 
-        fun bind(lifecycleOwner: LifecycleOwner, bookmark: BookmarkEntity, bookmarkViewModel: BookmarkViewModel) {
+        fun bind(lifecycleOwner: LifecycleOwner, bookmark: BookmarkEntity, homeViewModel: HomeViewModel) {
             itemBinding.setLifecycleOwner(lifecycleOwner)
-            //TODO: Databinding error
             itemBinding.tvBookmarkLocationContents.text = bookmark.address
             itemBinding.bookmark = bookmark
-            itemBinding.bookmarkViewModel = bookmarkViewModel
+            itemBinding.homeViewModel = homeViewModel
 
             Glide.with(itemBinding.root)
                 .load(bookmark.img_url)
