+++ app/src/main/res/layout/activity_home.xml
@@ -5,9 +5,6 @@
         tools:context=".ui.main.HomeActivity">
     <data>
         <variable
-                name="bookmarkViewModel"
-                type="kr.co.connect.boostcamp.livewhere.ui.main.BookmarkViewModel"/>
-        <variable
                 name="searchViewModel"
                 type="kr.co.connect.boostcamp.livewhere.ui.main.SearchViewModel"/>
         <variable
