+++ app/src/main/res/layout/fragment_home.xml
@@ -7,9 +7,6 @@
         <variable
             name="homeViewModel"
             type="kr.co.connect.boostcamp.livewhere.ui.main.HomeViewModel"/>
-        <variable
-            name="bookmarkViewModel"
-            type="kr.co.connect.boostcamp.livewhere.ui.main.BookmarkViewModel"/>
     </data>
     <androidx.coordinatorlayout.widget.CoordinatorLayout
             android:id="@+id/cl_home_backdrop"
@@ -63,7 +60,6 @@
                 android:layout_gravity="bottom"
                 app:behavior_peekHeight="@dimen/dropbar_header"
                 app:layout_behavior="@string/bottom_sheet_behavior"
-                bind:bookmarkViewModel="@{bookmarkViewModel}"
                 bind:homeViewModel="@{homeViewModel}"
                 app:behavior_hideable="false"/>
     </androidx.coordinatorlayout.widget.CoordinatorLayout>
